fix: Bracket placement in guard clause
The check in `process_eight` protects against out of bound issues, however the closing bracket is placed in the wrong position. What this means is if you run the script with a Warband with the number of units equal to a boundry condition (e.g. Mollog's Mob with 4) you will get an index out of bounds exception.

This PR fixes the bracket palcement and allows successful processing of warbands with those number of units.